requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
pdfminer.six
python-docx
langchain
PdfReader   
PyPDF2
docx_parser
python-docx
transformers
sentencepiece
python-dotenv
langchain-community
langchain_core
streamlit
python-multipart
spacy
en_core_web_sm
langchain_ollama
jinja2
routes
langchain_groq 